Add a --reparse_actions command to the bill task to re-run the action regexes after we discover a pattern that should be applied to existing data
Before updating a file, ask the user to confirm on the console.

We can't re-fetch bill data from before the 113th Congress because it doesn't exist in the GPO bill status data, so it's easiest if we update the archival files in-place. This may also be a faster way to test regex changes on existing data.